Сделан рефакторинг. Переименован модуль
