+++ js/main.js
@@ -1,3 +1,4 @@
+import './utils.js';
 import {arrayLodging} from './data.js';
 
 arrayLodging(10);
